* lydiasevelt/master:
  Updating version number display on pages.
  Remote master setup now working by running snmpd on sge_master host so that the webserver can pull the load average information. Still requires webserver to be configured as a submit host.
  Adding -t qstat queries so that it also displays timing information so that submit time/start time can be displayed.
  Updates to README to include basic instructions for qstat_reduce config. Additional changes to working in phpqstat.conf
  Removed personal information.
  Adding respect for qstat_reduce, running old code if not yes, otherwise pulling from cached xml file. Also included the display of which project a job is assigned to.
  Adding code to respect qstat_reduce, running the old code if it is not yes, otherwise reading from the single cached xml file.
  Adding header to qstat_job, the rest is still as it was, calling qstat ondemand. I might come back for this one, but I doubt users will query individual jobs all that often, and pulling information from just a single job doesn't seem to put much load on sge_master, so I'll let this remain ondemand for now.
  Modified qhost to utilize the qstat_reduce header, and to respect qstat_reduce, running the old code if qstat_reduce is not yet, otherwise pulling from the cached xml file.
  Adding -ext to qstat to provide additional output into xml file. This is useful for determining which project or department a job belongs to.
  Updated phpqstat.conf to include LOAD_WAIT variable, and the descriptions of CACHE_TIME and QSTAT_REDUCE with notifications that they must be set in config.inc.php
  Adding header.php for qstat_reduce refresh execution and display.
  Adding config.inc.php for qstat_reduce mode php variables.
  Adding qinfo.sh for qstat_reduce mode.
  Add -ncb to qhost call to surpress the processor topology, which fixes the qhost.php (Host status) display page.
  Fixing path to source settings.sh to use SGE_CELL variable.
  Migration to HPCNow! GitHUB repo

+++ qhost.php
@@ -11,6 +11,7 @@
 <?php
 $owner  = $_GET['owner'];
 echo "<body><table align=center width=95% border=\"1\" cellpadding=\"0\" cellspacing=\"0\"><tbody>";
+include("header.php");
 echo "<tr><td><h1>PHPQstat</h1></td></tr>
       <tr><td CLASS=\"bottom\" align=center><a href='index.php'>Home</a> *  <a href=\"qhost.php?owner=$owner\">Hosts status</a> *  <a href=\"qstat.php?owner=$owner\">Queue status</a> * <a href=\"qstat_user.php?owner=$owner\">Jobs status ($owner)</a> * <a href=\"about.php?owner=$owner\">About PHPQstat</a></td></tr>";
 ?>
@@ -32,21 +33,27 @@ echo "<tr><td><h1>PHPQstat</h1></td></tr>
                 <td>swap_used</td>
                 </tr>
 <?php
-$password_length = 20;
+if ($qstat_reduce != "yes") {
 
-function make_seed() {
-  list($usec, $sec) = explode(' ', microtime());
-  return (float) $sec + ((float) $usec * 100000);
+	$password_length = 20;
+
+	function make_seed() {
+	  list($usec, $sec) = explode(' ', microtime());
+	  return (float) $sec + ((float) $usec * 100000);
+	}
+
+	srand(make_seed());
+
+    $token=uniqid('phpqstat_');
+
+	$out = exec("./qhostout /tmp/$token.xml");
+
+	//printf("System Output: $out\n"); 
+	$qhost = simplexml_load_file("/tmp/$token.xml");
+} else {
+	$qhost = simplexml_load_file("/tmp/qhost.xml");
 }
 
-srand(make_seed());
-
-$token=uniqid('phpqstat_');
-
-$out = exec("./qhostout /tmp/$token.xml");
-
-//printf("System Output: $out\n"); 
-$qhost = simplexml_load_file("/tmp/$token.xml");
 $i=0;
 foreach ($qhost->host as $host) {
 	echo "<tr>";
@@ -59,8 +66,9 @@ foreach ($qhost->host as $host) {
 	$i++;
 }
 
-
-unlink("/tmp/$token.xml");
+if ($qstat_reduce != "yes") {
+    unlink("/tmp/$token.xml");
+}
 
 ?>
